Added check of device to see if it is a symbolic link. Moved some stuff in order to avoid a Segmentation fault in some cases when calling SIGINT/SIGKILL signal.
git-svn-id: https://csrg.inf.utfsm.cl/repos/software/CSAT@756 75a0fe40-ac5b-440f-90d5-6af9f3fbf44f

+++ extClients/src/Telescope.cpp
@@ -5,18 +5,32 @@ Telescope::Telescope(bool isLocal)
 	this->isLocal = isLocal;
 	this->serial   = "/dev/ttyS0";
 	this->serialbk = "/dev/ttyS0.bk";
+	this->run = true;
+	this->move = false;
+	this->csatC = new CSATClient();
+	this->cscRun = false;
+	this->cssRun = false;
+	this->fdm = 0;
+	this->fds = 0;
 }
 
 Telescope::~Telescope()
 {
+	this->csatC->stop();
+	delete this->csatC;
 	if( this->isLocal ) {
-		close(this->fdm);
-		close(this->fds);
-		unlink(this->serial);
-		rename(this->serialbk,serial);
+		if(this->fdm > 0)
+			close(this->fdm);
+		if(this->fds > 0)
+			close(this->fds);
+		if(this->move){
+			unlink(this->serial);
+			rename(this->serialbk,serial);
+		}
 	}
 	else {
-		close(this->fdm);
+		if(this->fdm > 0)
+			close(this->fdm);
 	}
 }
 
@@ -27,7 +41,7 @@ void Telescope::parseInstructions()
 int Telescope::start()
 {
 	char *slavename;
-	//char buf;
+	struct stat buf;
 	static struct termios termorig;  /* static for zero's */
 
 	if( this->isLocal ) {
@@ -38,18 +52,20 @@ int Telescope::start()
 		unlockpt(this->fdm); // unlock slave
 		setsid();
 		if( (slavename = ptsname(this->fdm)) == NULL) // get name of slave
-		{
-			close(this->fdm);
+			return -1;
+		lstat(this->serial, &buf);
+		if ((buf.st_mode & S_IFMT) == S_IFLNK){
+			printf("Device is a Symlink, check it.\n");
 			return -1;
 		}
 		if( (this->fds = open(slavename, O_RDWR)) < 0) // open slave
-		{
-			close(this->fdm);
 			return -1;
+		if(!strcmp(this->serial,slavename)){
+			rename(this->serial,this->serialbk);
+			this->move = true;
+			if(symlink(slavename,this->serial)!=0)
+				printf("Unable to create Symlink\n");
 		}
-		rename(this->serial,this->serialbk);
-		if(symlink(slavename,this->serial)!=0)
-			printf("Unable\n");
 		chmod(slavename, S_IRWXU|S_IRWXG|S_IRWXO);
 		printf("fds: %d\n", this->fds);
 		ioctl(this->fds, I_PUSH, "ptem"); // push ptem
@@ -61,17 +77,25 @@ int Telescope::start()
 		termorig.c_lflag = 0;
 		(void) ioctl(this->fds, TCSETS, &termorig);
 	}
-
 	else {
 		slavename = (char *)this->serial;
 		this->fdm = open(slavename,O_RDWR);
 		this->configPort();
 	}
-
 	printf("%s\n\n", slavename);
-
+	this->cscRun = (this->csatC->startCSC() == 0);
+	this->cssRun = (this->csatC->startCSS() == 0);
 	this->parseInstructions();
 	return 0;
+}
+
+void Telescope::stop() {
+	char buf = '#';
+	this->run = false;
+	if(this->isLocal){
+		write(this->fds,&buf,1);
+		write(this->fds,&buf,1);
+	}
 }
 
 void Telescope::configPort() {
